### 2024-02-25 #6408 @danielskatz
@nathanReitinger - since you are the only author, you don't need the line ```    equal-contrib: true``` in your paper.  Please remove this, then use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.
