### 2024-02-25 #6408 @danielskatz
👋 @sneakers-the-rat - would you be willing to edit this JOSS submission?
