### 2024-02-26 3b5ebd9
paper bib DOI and contrib-fix
