+++ paper/paper.bib
@@ -5,7 +5,9 @@
   volume={22},
   pages={1},
   year={2019},
-  publisher={HeinOnline}
+  publisher={HeinOnline}, 
+  doi = {10.31228/osf.io/bfqh3},
+  url = {https://doi.org/10.31228/osf.io/bfqh3}
 }
 
 @article{reitinger:2023,
@@ -20,13 +22,15 @@
 }
 
 @article{reitinger:2021,
-  title={ML-CB: Machine Learning Canvas Block},
+  title={ML-CB: Machine learning canvas block},
   author={Reitinger, Nathan and Mazurek, Michelle L},
   journal={Proc. Priv. Enhancing Technol.},
   volume={2021},
   number={3},
   pages={453--473},
-  year={2021}
+  year={2021}, 
+  doi = {10.2478/popets-2021-0056},
+  url = {https://doi.org/10.2478/popets-2021-0056}
 }
 
 @inproceedings{thomas:2021,
@@ -35,7 +39,9 @@
   booktitle={2021 IEEE Symposium on Security and Privacy (SP)},
   pages={247--267},
   year={2021},
-  organization={IEEE}
+  organization={IEEE}, 
+  doi = {10.1109/sp40001.2021.00028},
+  url = {https://doi.org/10.1109/sp40001.2021.00028}
 }
 
 @inproceedings{singhal:2023,
@@ -44,19 +50,23 @@
   booktitle={2023 IEEE 8th European Symposium on Security and Privacy (EuroS\&P)},
   pages={868--895},
   year={2023},
-  organization={IEEE}
+  organization={IEEE}, 
+  doi = {10.1109/eurosp57164.2023.00056},
+  url = {https://doi.org/10.1109/eurosp57164.2023.00056}
 }
 
 @inproceedings{franz:2021,
-  title={$\{$SoK$\}$: Still Plenty of Phish in the Sea—A Taxonomy of $\{$User-Oriented$\}$ Phishing Interventions and Avenues for Future Research},
+  title={SoK: Still plenty of phish in the sea—A taxonomy of user-oriented phishing interventions and avenues for future research},
   author={Franz, Anjuli and Zimmermann, Verena and Albrecht, Gregor and Hartwig, Katrin and Reuter, Christian and Benlian, Alexander and Vogt, Joachim},
   booktitle={Seventeenth Symposium on Usable Privacy and Security (SOUPS 2021)},
   pages={339--358},
-  year={2021}
+  year={2021}, 
+  doi = {10.26083/tuprints-00020675},
+  url = {https://doi.org/10.26083/tuprints-00020675}
 }
 
 @article{reitinger:sok,
-  title={SoK: Considerations in Measuring Compliance with Privacy Regulations},
+  title={SoK: Considerations in measuring compliance with privacy regulations},
   author={Reitinger, Nathan and Mazurek, Michelle L}
 }
 
