### 2024-02-27 #6408 @sneakers-the-rat
Still working on editing my first review and also handling one for pyopensci at the moment, if it's alright i'll get the next one, though this does look very interesting!
