### 2024-03-07 #6408 @Nikoleta-v3
Hello @nathanReitinger! Thank you again for your submission. At first glance, the paper and package look very interesting 🚀 🙌🏻

I have a few comments at this point 😄

- The README of the repository is quite poor. It's hard to understand the functionality of the package or find any details without going to the online documentation, which is great, by the way! I recommend that you might want to add a bit to the README, maybe even some screenshots from the demo video to showcase the package and its functionality.

- Another minor comment is about [testing section](https://nathanreitinger.github.io/CART/development/): could you add some details on how to run the tests?

I believe that there are many people who can review this package, so I am not worried about finding reviewers. The list of potential reviewers is here: https://reviewers.joss.theoj.org/lookup. 😄 Do you have any preferences?
