+++ requirements.txt
@@ -1,11 +1,34 @@
 # Automatically generated by https://github.com/damnever/pigar.
 
-filelock==3.4.0
-Flask==2.2.5
-Flask-Session==0.5.0
-moment==0.12.1
-ngrok-flask-cart==0.0.7
-numpy==1.21.6
-pandas==1.3.3
-plotly==5.18.0
-selenium==4.11.2
+# CART/cart.py: 13
+Flask == 2.2.5
+
+# CART/cart.py: 23
+filelock == 3.4.0
+
+# CART/cart.py: 21
+flask_session == 0.5.0
+
+# CART/cart.py: 31
+moment == 0.12.1
+
+# CART/cart.py: 38
+ngrok_flask_cart == 0.0.1
+
+# CART/cart.py: 27
+# CART/sample_get_data/get_data.py: 19
+numpy == 1.21.6
+
+# CART/cart.py: 26
+# CART/sample_get_data/get_data.py: 11
+pandas == 1.3.3
+
+# CART/cart.py: 57,58
+plotly == 5.18.0
+
+# CART/sample_get_data/get_data.py: 2,3,4,5,6,7,8,9
+# CART/testing/test.py: 2,4,5,6,7,8,9,10
+selenium == 3.141.0
+
+# CART/testing/test.py: 3
+webdriver_manager == 4.0.0
