+++ docs/development.md
@@ -12,14 +12,33 @@ Please note that this project is released with a [Contributor Code of Conduct][c
 
 ## Areas for improvement
 
-| Area   | Comment                                                      |
-| ------ | ------------------------------------------------------------ |
-| Mobile | `CART` currently only works for desktop.                     |
-| GUI    | Running `cart.py` may benefit from a GUI that allows users to select options instead of passing in flags. |
-| Login  | `CART` assumes users are honest and therefore does not require login passwords or other security-facing features. Some teams may prefer tooling that helps enforce access control. |
+| Area            | Comment                                                      |
+| --------------- | ------------------------------------------------------------ |
+| Mobile          | `CART` currently only works for desktop.                     |
+| GUI             | Running `cart.py` may benefit from a GUI that allows users to select options instead of passing in flags. |
+| Login           | `CART` assumes users are honest and therefore does not require login passwords or other security-facing features. Some teams may prefer tooling that helps enforce access control, like logins with a password. |
+| Post-processing | `CART` would benefit from a task that compiles all votes cast into a single csv file; currently, `CART` stores papers (as `.csv` files) with votes cast per user, but does not aggregate all of these into a single file for completion post-processing. |
 
 
 
 ## Testing 
 
-Testing scripts are found in the `/testing` folder. 
+Testing scripts are found in the `/testing` folder. The `test.py` program will
+
+- use Selenium for testing
+- setup `CART` by moving test abstracts into the right folder
+  - abstracts will be moved into the `abstracts/-testing/` folder (and cleaned up when `test.py` exits) 
+- use a "small" set of abstracts
+  - login as a provided user and vote on 5 abstracts (`number_to_vote` argument)
+    - checking login 
+    - checking vote casting
+    - checking "done" or completed on all abstracts voted on
+    - teardown to reset abstracts
+- use a "big" set of abstracts 
+  - login as a provided user and vote on 50 abstracts (`number_to_vote` argument)
+    - checking login 
+    - checking vote casting
+    - checking "done" or completed on all abstracts voted on
+    - teardown to reset abstracts
+- print results of tests to terminal 
+
