### 2024-03-14 #6408 @nathanReitinger
Thanks for the comments @Nikoleta-v3 🎉🎉🎉
- the [readme](https://github.com/nathanReitinger/CART?tab=readme-ov-file#overview) was updated to include basic information about the package
- additional information on [testing](https://nathanreitinger.github.io/CART/development/#testing) was included in the docs

For reviewer preference, I have none. I was trying to see if there are any CS-python-javascript reviewers, but there are quite a few ([reviewer search](https://reviewers.joss.theoj.org/reviewers/search?language=26&area_id=&keywords=computer+science)) so I don't have a preference.
