### 2024-03-14 #6408 @Nikoleta-v3
@emmagbugba @anthony-arnold @atanikan @padhoni would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "CART: A Tool for Making Paper Relevancy Screening Easier". You can find more information at the top of this Github issue 🆙
