### 2024-03-14 #6408 @padhoni
Sure, will do.

Thanks and Regards,

Pan Dhoni
602-748-9515

On Thu, Mar 14, 2024 at 3:35 PM Nikoleta Glynatsi ***@***.***>
wrote:

> @emmagbugba <https://github.com/emmagbugba> @anthony-arnold
> <https://github.com/anthony-arnold> @atanikan
> <https://github.com/atanikan> @padhoni <https://github.com/padhoni> would
> any of you be willing to review this submission for JOSS?
>
> We carry out our checklist-driven reviews here in GitHub issues and follow
> these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
> The submission I'd like you to review is titled: "CART: A Tool for Making
> Paper Relevancy Screening Easier". You can find more information at the top
> of this Github issue 🆙
>
> Please let me know if you're available 😄 Thank you!
>
