### 2024-03-16 #6494 @editorialbot
[REVIEW]: CART: A Tool for Making Paper Relevancy Screening Easier
**Submitting author:** @nathanReitinger (<a href="http://orcid.org/0000-0001-8294-9946">Nathan Reitinger</a>)
**Repository:** https://github.com/nathanReitinger/CART
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
