### 2024-03-16 #6494 @Nikoleta-v3
Could please take care of this: https://github.com/openjournals/joss-reviews/issues/6494#issuecomment-2001939714?
