+++ paper/paper.bib
@@ -18,10 +18,11 @@
   number={3},
   pages={263--317},
   year={2023},
+  xdoi = {},
   publisher={HeinOnline}
 }
 
-@article{reitinger:2021,
+@inproceedings{reitinger:2021,
   title={ML-CB: Machine learning canvas block},
   author={Reitinger, Nathan and Mazurek, Michelle L},
   journal={Proc. Priv. Enhancing Technol.},
@@ -67,26 +68,28 @@
 
 @article{reitinger:sok,
   title={SoK: Considerations in measuring compliance with privacy regulations},
+  xdoi = {},
   author={Reitinger, Nathan and Mazurek, Michelle L}
 }
 
-@article{smit:2002,
-  title={Atlas. ti for qualitative data analysis},
-  author={Smit, Brigitte},
-  journal={Perspectives in education},
+@article{smit:2021,
+  title={Computer-Assisted Qualitative Data Analysis Software for Scoping Reviews: A Case of ATLAS.ti},
+  author={Smit, Brigitte and Scherman, Vanessa},
+  journal={International Journal of Qualitative Methods},
   volume={20},
-  number={3},
-  pages={65--75},
-  year={2002},
-  publisher={University of the Free State}
+  pages={1--3},
+  year={2021},
+  doi = {10.1177/16094069211019140},
+  publisher={SAGE Publications}
 }
 
-@article{godau:2003,
-  title={Qualitative data analysis software: MAXqda},
-  author={Godau, RI},
-  journal={Qualitative Research Journal},
-  volume={4},
-  number={1},
-  pages={66--72},
-  year={2003}
+@article{Franzosi:2012,
+  title={Quantitative narrative analysis software options compared: PC-ACE and CAQDAS (ATLAS.ti, MAXqda, and NVivo)},
+  author={Franzosi, Roberto and Doyle, Sophie and McClelland, Laura E and Putnam Rankin, Caddie and Vicari, Stefania},
+  journal={Quality \& Quantity},
+  volume={47},
+  pages={3219--3247},
+  year={2013},
+  doi = {10.1007/s11135-012-9714-3},
+  publisher={Springer}
 }
