+++ paper/paper.md
@@ -32,7 +32,7 @@ bibliography: paper.bib
 
 A systematization of knowledge or systematic review paper is a popular class of paper found across multiple disciplines: medicine, computer science, sociology, psychology, economics, political science, marketing,  and genetics. This type of paper most commonly involves a two-stage process. In the first stage, researchers gather a large set of papers—easily over 10,000—and then manually filter those papers (based on paper titles and abstracts) according to predefined relevancy criteria [@thomas:2021; @singhal:2023; @franz:2021; @reitinger:sok]. For example, if a researcher wanted to learn about legally-sufficient data sanitization tools [@bellovin:2019], a paper on the prevalence of a particular web tracking technique like canvas fingerprinting [@reitinger:2021] would be marked as irrelevant, but a paper on statutory differentially privacy requirements [@reitinger:2023] would be marked as relevant. After this first step is complete—often resulting in a set of around 100 papers—most researchers move on to analyzing papers in-depth, using qualitative coding techniques well-fit by existing software.
 
-Surprisingly, software to support this first step of paper culling is currently lacking. Existing tools are either too generalist (resulting in ineffective paper reviewing) or too fine-tuned (resulting in tradeoffs impacting functionality). For example, generalist tools like Microsoft Excel or Google Sheets could allow teams of researchers to check the relevancy of snippets of text, but these tools can be clunky (i.e., given the scale of papers in consideration) and not usable (i.e., reviewing hundreds of rows in a spreadsheet is error-prone and fatiguing). Other tools, like Atlas.ti [@smit:2002] or maxQDA [@godau:2003], could allow teams of researchers to "code" for relevancy, but do not naturally accommodate dynamic participation (e.g., team members might not be able to review the same amount of papers) or variable review rates (e.g., sometimes you want all team members to review all papers, and sometimes you want at least two, but you do not set who those team members are). 
+Surprisingly, software to support this first step of paper culling is currently lacking. Existing tools are either too generalist (resulting in ineffective paper reviewing) or too fine-tuned (resulting in tradeoffs impacting functionality). For example, generalist tools like Microsoft Excel or Google Sheets could allow teams of researchers to check the relevancy of snippets of text, but these tools can be clunky (i.e., given the scale of papers in consideration) and not usable (i.e., reviewing hundreds of rows in a spreadsheet is error-prone and fatiguing). Other tools, like Atlas.ti [@smit:2021] or maxQDA [@Franzosi:2012], could allow teams of researchers to "code" for relevancy, but do not naturally accommodate dynamic participation (e.g., team members might not be able to review the same amount of papers) or variable review rates (e.g., sometimes you want all team members to review all papers, and sometimes you want at least two, but you do not set who those team members are). 
 
 
 
@@ -44,11 +44,11 @@ Surprisingly, software to support this first step of paper culling is currently 
 
 
 
-![Architecture of `CART`.\label{fig:arch}](docs/img/architecture.svg)
+![Architecture of `CART`.\label{fig:arch}](../docs/img/architecture.svg)
 
 *First*, `CART` runs a Flask application responsible for serving and updating papers (i.e., represented as an abstract, title, URL, and a few metadata fields). A team member will only be served an abstract that they have not seen before, needs votes, has not been flagged as "not" a research paper (i.e., a proceedings or similar document), and is not being viewed by another user. After a vote is cast, the Flask application will update the paper to log who voted, what their vote was, and how many votes have been cast for the paper. This process of serving and updating papers will loop until there are no more eligible papers available. 
 
-![Web interface tabs: home, progress, history, information, and account.\label{fig:tabs}](docs/img/dock.svg)
[…]
