### 2024-03-19 8a644c2
Delete paper.bib
