+++ paper.bib
@@ -1,82 +1,0 @@
-@article{bellovin:2019,
-  title={Privacy and synthetic datasets},
-  author={Bellovin, Steven M and Dutta, Preetam K and Reitinger, Nathan},
-  journal={Stan. Tech. L. Rev.},
-  volume={22},
-  pages={1},
-  year={2019},
-  publisher={HeinOnline}
-}
-
-@article{reitinger:2023,
-  title={Epsilon-differential privacy, and a twostep test for quantifying reidentification risk},
-  author={Reitinger, Nathan and Deshpande, Amol},
-  journal={Jurimetrics: The Journal of Law, Science \& Technology},
-  volume={63},
-  number={3},
-  pages={263--317},
-  year={2023},
-  publisher={HeinOnline}
-}
-
-@article{reitinger:2021,
-  title={ML-CB: Machine Learning Canvas Block},
-  author={Reitinger, Nathan and Mazurek, Michelle L},
-  journal={Proc. Priv. Enhancing Technol.},
-  volume={2021},
-  number={3},
-  pages={453--473},
-  year={2021}
-}
-
-@inproceedings{thomas:2021,
-  title={Sok: Hate, harassment, and the changing landscape of online abuse},
-  author={Thomas, Kurt and Akhawe, Devdatta and Bailey, Michael and Boneh, Dan and Bursztein, Elie and Consolvo, Sunny and Dell, Nicola and Durumeric, Zakir and Kelley, Patrick Gage and Kumar, Deepak and others},
-  booktitle={2021 IEEE Symposium on Security and Privacy (SP)},
-  pages={247--267},
-  year={2021},
-  organization={IEEE}
-}
-
-@inproceedings{singhal:2023,
-  title={SoK: Content moderation in social media, from guidelines to enforcement, and research to practice},
-  author={Singhal, Mohit and Ling, Chen and Paudel, Pujan and Thota, Poojitha and Kumarswamy, Nihal and Stringhini, Gianluca and Nilizadeh, Shirin},
-  booktitle={2023 IEEE 8th European Symposium on Security and Privacy (EuroS\&P)},
-  pages={868--895},
-  year={2023},
-  organization={IEEE}
-}
-
-@inproceedings{franz:2021,
-  title={$\{$SoK$\}$: Still Plenty of Phish in the Sea—A Taxonomy of $\{$User-Oriented$\}$ Phishing Interventions and Avenues for Future Research},
-  author={Franz, Anjuli and Zimmermann, Verena and Albrecht, Gregor and Hartwig, Katrin and Reuter, Christian and Benlian, Alexander and Vogt, Joachim},
-  booktitle={Seventeenth Symposium on Usable Privacy and Security (SOUPS 2021)},
-  pages={339--358},
-  year={2021}
-}
-
-@article{reitinger:sok,
-  title={SoK: Considerations in Measuring Compliance with Privacy Regulations},
-  author={Reitinger, Nathan and Mazurek, Michelle L}
-}
-
-@article{smit:2002,
-  title={Atlas. ti for qualitative data analysis},
-  author={Smit, Brigitte},
-  journal={Perspectives in education},
-  volume={20},
-  number={3},
-  pages={65--75},
-  year={2002},
-  publisher={University of the Free State}
-}
-
-@article{godau:2003,
-  title={Qualitative data analysis software: MAXqda},
-  author={Godau, RI},
-  journal={Qualitative Research Journal},
-  volume={4},
-  number={1},
-  pages={66--72},
-  year={2003}
-}
