### 2024-03-19 19b962f
Delete paper.md
