+++ paper/paper.bib
@@ -18,6 +18,7 @@
   number={3},
   pages={263--317},
   year={2023},
+  url = {https://www.americanbar.org/content/dam/aba/publications/Jurimetrics/spring-2023/epsilon-differential-privacy-and-a-two-step-test-for-quantifying-reidentification-risk.pdf},
   xdoi = {},
   publisher={HeinOnline}
 }
@@ -66,9 +67,12 @@
   url = {https://doi.org/10.26083/tuprints-00020675}
 }
 
-@article{reitinger:sok,
+@inproceedings{reitinger:sok,
   title={SoK: Considerations in measuring compliance with privacy regulations},
   xdoi = {},
+  booktitle={7th Workshop on Technology and Consumer Protection (ConPro 2023)},
+  year={2023}, 
+  url = {https://conpro23.ieee-security.org/papers/reitinger-conpro23.pdf},
   author={Reitinger, Nathan and Mazurek, Michelle L}
 }
 
