### 2024-04-03 #6494 @padhoni
Hi All- Apologies for the late reply, was outside due to personal reasons,
now I have time and will contribute from tomorrow. Let me go through email
thread and set up related accounts.

Thanks and Regards,

Pan Dhoni
602-748-9515

On Sat, Mar 16, 2024 at 6:18 AM The Open Journals editorial robot <
***@***.***> wrote:

> *Submitting author:* @nathanReitinger <https://github.com/nathanReitinger>
> (Nathan Reitinger <http://orcid.org/0000-0001-8294-9946>)
> *Repository:* https://github.com/nathanReitinger/CART
> *Branch with paper.md* (empty if default branch):
> *Version:* v1.0.0
> *Editor:* @Nikoleta-v3 <https://github.com/Nikoleta-v3>
> *Reviewers:* @padhoni <https://github.com/padhoni>, @atanikan
> <https://github.com/atanikan>
> *Archive:* Pending
> Status
>
> [image: status]
> <https://joss.theoj.org/papers/d58ad58b740d1d2d2470beaf533ef221>
>
> Status badge code:
>
>
> *Reviewers and authors*:
>
> Please avoid lengthy details of difficulties in the review thread.
> Instead, please create a new issue in the target repository
> <https://github.com/nathanReitinger/CART> and link to those issues
> (especially acceptance-blockers) by leaving comments in the review thread
> below. (For completists: if the target issue tracker is also on GitHub,
> linking the review thread in the issue or vice versa will create
> corresponding breadcrumb trails in the link target.)
> Reviewer instructions & questions
>
> @padhoni <https://github.com/padhoni> & @atanikan
> <https://github.com/atanikan>, your review will be checklist based. Each
> of you will have a separate checklist that you should update when carrying
> out your review.
> First of all you need to run this command in a separate comment to create
> the checklist:
>
> @editorialbot generate my checklist
>
> The reviewer guidelines are available here:
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Any
> questions/concerns please let @Nikoleta-v3
> <https://github.com/Nikoleta-v3> know.
>
> ✨ *Please start on your review when you are able, and be sure to complete
> your review in the next six weeks, at the very latest* ✨
> Checklists
>
> @padhoni <https://github.com/padhoni>, please create your checklist
> typing: @editorialbot generate my checklist
>
> @atanikan <https://github.com/atanikan>, please create your checklist
> typing: @editorialbot generate my checklist
>
