### 2024-06-04 nathanReitinger/CART#1 @atanikan
Updates to installation steps
I am the reviewer for JOSS [#6494](https://github.com/openjournals/joss-reviews/issues/6494)
I would recommend the following updates to the installation steps:
a) Use a virtual environment like python3 -m venv or conda create -n ...
b) Remove the need to be a sudo user.
c) Mention the prerequisite of the Python version needed to make the installation work. If not, with the Python version of 3.11.6 I see the following error

```bash
(CART_REVIEW) alcfwl184:CART adityatanikanti$ pip3 install -r requirements.txt
Collecting Flask==2.2.5 (from -r requirements.txt (line 4))
  Obtaining dependency information for Flask==2.2.5 from https://files.pythonhosted.org/packages/9f/1a/8b6d48162861009d1e017a9740431c78d860809773b66cac220a11aa3310/Flask-2.2.5-py3-none-any.whl.metadata
  Downloading Flask-2.2.5-py3-none-any.whl.metadata (3.9 kB)
Collecting filelock==3.4.0 (from -r requirements.txt (line 7))
  Obtaining dependency information for filelock==3.4.0 from https://files.pythonhosted.org/packages/e8/3b/b59c7bcacc4cccafcd6e927a9e191268657e79a0a75530132cbf03b22c47/filelock-3.4.0-py3-none-any.whl.metadata
  Downloading filelock-3.4.0-py3-none-any.whl.metadata (2.6 kB)
Collecting flask_session==0.5.0 (from -r requirements.txt (line 10))
  Obtaining dependency information for flask_session==0.5.0 from https://files.pythonhosted.org/packages/3e/d4/b374183251054928ddb5e72f9a0d3d764d0f2af0638fbbdf205df26e55e3/flask_session-0.5.0-py3-none-any.whl.metadata
  Downloading flask_session-0.5.0-py3-none-any.whl.metadata (1.4 kB)
Collecting moment==0.12.1 (from -r requirements.txt (line 13))
  Obtaining dependency information for moment==0.12.1 from https://files.pythonhosted.org/packages/15/a7/052de228d843337a5eec8fb7e368fe08daaa66f2679354163b77d416ef0b/moment-0.12.1-py3-none-any.whl.metadata
  Downloading moment-0.12.1-py3-none-any.whl.metadata (711 bytes)
Collecting ngrok_flask_cart==0.0.1 (from -r requirements.txt (line 16))
  Downloading ngrok-flask-cart-0.0.1.tar.gz (3.1 kB)
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
ERROR: Ignored the following versions that require a different python version: 1.21.2 Requires-Python >=3.7,<3.11; 1.21.3 Requires-Python >=3.7,<3.11; 1.21.4 Requires-Python >=3.7,<3.11; 1.21.5 Requires-Python >=3.7,<3.11; 1.21.6 Requires-Python >=3.7,<3.11
ERROR: Could not find a version that satisfies the requirement numpy==1.21.6 (from versions: 1.3.0, 1.4.1, 1.5.0, 1.5.1, 1.6.0, 1.6.1, 1.6.2, 1.7.0, 1.7.1, 1.7.2, 1.8.0, 1.8.1, 1.8.2, 1.9.0, 1.9.1, 1.9.2, 1.9.3, 1.10.0.post2, 1.10.1, 1.10.2, 1.10.4, 1.11.0, 1.11.1, 1.11.2, 1.11.3, 1.12.0, 1.12.1, 1.13.0, 1.13.1, 1.13.3, 1.14.0, 1.14.1, 1.14.2, 1.14.3, 1.14.4, 1.14.5, 1.14.6, 1.15.0, 1.15.1, 1.15.2, 1.15.3, 1.15.4, 1.16.0, 1.16.1, 1.16.2, 1.16.3, 1.16.4, 1.16.5, 1.16.6, 1.17.0, 1.17.1, 1.17.2, 1.17.3, 1.17.4, 1.17.5, 1.18.0, 1.18.1, 1.18.2, 1.18.3, 1.18.4, 1.18.5, 1.19.0, 1.19.1, 1.19.2, 1.19.3, 1.19.4, 1.19.5, 1.20.0, 1.20.1, 1.20.2, 1.20.3, 1.21.0, 1.21.1, 1.22.0, 1.22.1, 1.22.2, 1.22.3, 1.22.4, 1.23.0rc1, 1.23.0rc2, 1.23.0rc3, 1.23.0, 1.23.1, 1.23.2, 1.23.3, 1.23.4, 1.23.5, 1.24.0rc1, 1.24.0rc2, 1.24.0, 1.24.1, 1.24.2, 1.24.3, 1.24.4, 1.25.0rc1, 1.25.0, 1.25.1, 1.25.2, 1.26.0b1, 1.26.0rc1, 1.26.0, 1.26.1, 1.26.2, 1.26.3, 1.26.4, 2.0.0b1, 2.0.0rc1, 2.0.0rc2)
ERROR: No matching distribution found for numpy==1.21.6

[notice] A new release of pip is available: 23.2.1 -> 24.0
[notice] To update, run: pip3 install --upgrade pip
(CART_REVIEW) alcfwl184:CART adityatanikanti$ python3 --version
Python 3.11.6
```
d) This optional but it'd help if you can provide a container-based (like docker) solution for installation as an alternative
