### 2024-06-04 nathanReitinger/CART#2 @atanikan
Read the docs link (to self) is broken
[#6494](https://github.com/openjournals/joss-reviews/issues/6494) . The Read the docs link (to self) is broken here https://nathanreitinger.github.io/CART/
