+++ cart.yml
@@ -1,0 +1,85 @@
+name: cart
+channels:
+  - conda-forge
+  - defaults
+dependencies:
+  - bzip2=1.0.8=h0d85af4_4
+  - ca-certificates=2022.9.24=h033912b_0
+  - libffi=3.4.2=h0d85af4_5
+  - libsqlite=3.40.0=ha978bb4_0
+  - libzlib=1.2.13=hfd90126_4
+  - ncurses=6.3=h96cf925_1
+  - openssl=3.0.7=hfd90126_0
+  - pip=22.3.1=pyhd8ed1ab_0
+  - python=3.10.0=h38b4d05_3_cpython
+  - readline=8.1.2=h3899abd_0
+  - setuptools=65.5.1=pyhd8ed1ab_0
+  - sqlite=3.40.0=h9ae0607_0
+  - tk=8.6.12=h5dbffcc_0
+  - wheel=0.38.4=pyhd8ed1ab_0
+  - xz=5.2.6=h775f41a_0
+  - pip:
+      - aiohttp==3.9.5
+      - aiosignal==1.3.1
+      - arrow==1.3.0
+      - async-timeout==4.0.3
+      - attrs==23.2.0
+      - blinker==1.8.2
+      - cachelib==0.13.0
+      - certifi==2024.6.2
+      - charset-normalizer==3.3.2
+      - click==8.1.7
+      - dateparser==1.2.0
+      - exceptiongroup==1.2.1
+      - fastjsonschema==2.19.1
+      - filelock==3.14.0
+      - flask==3.0.3
+      - flask-session==0.8.0
+      - frozenlist==1.4.1
+      - h11==0.14.0
+      - idna==3.7
+      - itsdangerous==2.2.0
+      - jinja2==3.1.4
+      - jsonschema==4.22.0
+      - jsonschema-specifications==2023.12.1
+      - jupyter-core==5.7.2
+      - markupsafe==2.1.5
+      - moment==0.12.1
+      - msgspec==0.18.6
+      - multidict==6.0.5
+      - nbformat==5.10.4
+      - ngrok-flask-cart==0.0.7
+      - numpy==1.26.4
+      - outcome==1.3.0.post0
+      - packaging==24.0
+      - pandas==2.2.2
+      - pigar==2.1.4
+      - platformdirs==4.2.2
+      - plotly==5.22.0
+      - pysocks==1.7.1
+      - python-dateutil==2.9.0.post0
+      - python-dotenv==1.0.1
+      - pytz==2024.1
+      - referencing==0.35.1
+      - regex==2024.5.15
+      - requests==2.32.3
+      - rpds-py==0.18.1
+      - selenium==4.21.0
+      - six==1.16.0
+      - sniffio==1.3.1
+      - sortedcontainers==2.4.0
+      - tenacity==8.3.0
+      - times==0.7
+      - traitlets==5.14.3
+      - trio==0.25.1
+      - trio-websocket==0.11.1
+      - types-python-dateutil==2.9.0.20240316
+      - typing-extensions==4.12.1
+      - tzdata==2024.1
+      - tzlocal==5.2
+      - urllib3==2.2.1
+      - webdriver-manager==4.0.1
+      - werkzeug==3.0.3
+      - wsproto==1.2.0
+      - yarl==1.9.4
+prefix: /usr/local/anaconda3/envs/cart
