### 2024-06-05 nathanReitinger/CART#1 @nathanReitinger
Thanks for the notes @atanikan, wonderful suggestions. Please pull the latest version to see these changes!

a) done, added `cart.yml` which you can use via `conda env create -f cart.yml`
b) done, not sure why I included `sudo` in the first place, but it seems it wasn't needed
c) I think that error was related to numpy. I tested on Python 3.11.0 and Python 3.10.0 and everything seemed to install correctly, though possibly there is an update with the minor update in 3.11.6. I think this new conda environment will fix the issue, but report back here if not.
d) let's see if there are more installation issues and, if so, I'll check out wrapping this in docker.
