### 2024-06-18 #6494 @jhculb
Hi @nathanReitinger, hope you're well :)

Functionality - Installation:
I've got a small problem - the path "abstracts/-sample_from_scrape/SAMPLE_PETS_at_Thu Feb 15 14:32:05 2024" is an invalid windows path, so cloning the repo would be possible but checking out would fail. Could this be fixed for windows users please?

Also could I check who the other contributor ([temp-9384289](https://github.com/temp-9384289)) is to this repo please? - I don't see any activity so could understand if this is an artifact somewhere. :)
