+++ paper/paper.md
@@ -48,7 +48,7 @@ Surprisingly, software to support this first step of paper culling is currently 
 
 *First*, `CART` runs a Flask application responsible for serving and updating papers (i.e., represented as an abstract, title, URL, and a few metadata fields). A team member will only be served an abstract that they have not seen before, needs votes, has not been flagged as "not" a research paper (i.e., a proceedings or similar document), and is not being viewed by another user. After a vote is cast, the Flask application will update the paper to log who voted, what their vote was, and how many votes have been cast for the paper. This process of serving and updating papers will loop until there are no more eligible papers available. 
 
-![Web interface tabs: home, progress, history, information, and account.\label{fig:tabs}](../docs/img/dock.svg)
+![Web interface tabs: home, progress, history, information, and account.\label{fig:tabs}](../docs/img/dock.png)
 
 *Second*, `CART` creates a gamified user experience for reviewing papers. The web interface makes voting on a paper simple (i.e., button click or keyboard press), allows users to set personal goals, and rewards (with confetti) users on every 50th (or set by the team) vote submitted. The web interface also includes tabs (\autoref{fig:tabs}) for **voting** on papers (home), viewing the team's **progress** (bar chart, total vote counts), viewing personal voting **history** (last 50 edits, permitting a change-of-vote), logging information **about** the project (collaborative editing with backups, where users may write guidelines for voting), and viewing **account** details (logging in and out). Moreover, the web app is optimized to provide information in a timely fashion (e.g., storing 'open' papers in logs rather than reviewing the full corpus of possible papers to find open papers). 
 
