### 2024-06-19 #6494 @nathanReitinger
> Functionality - Installation: I've got a small problem - the path "abstracts/-sample_from_scrape/SAMPLE_PETS_at_Thu Feb 15 14:32:05 2024" is an invalid windows path, so cloning the repo would be possible but checking out would fail. Could this be fixed for windows users please?

- I changed the name to `SAMPLE_PETS_feb152024` thinking/hoping the colon or slashes caused the issue. Please let me know if this doesn't fix it!

> Also could I check who the other contributor ([temp-9384289](https://github.com/temp-9384289)) is to this repo please?  I don't see any activity so could understand if this is an artifact somewhere. :)

- Actually, that account is mine, from a different computer. I had it created to make an anonymous GitHub repo for a separate project. Should I try and change the name or is there a good way to handle this?
