### 2024-06-26 #6494 @jhculb
Hi Nathan, sorry for the delay - I nuked my system through another project and had to reinstall the OS.

I'm using a fresh install of conda and am getting a PackagesNotFoundError when running 'conda env create -f cart.yml'.

I've tried this with the 3.10 and 3.12 miniconda environments available here: https://docs.anaconda.com/miniconda/miniconda-other-installer-links/

Can you help with this one by refining the install instructions?
--

On your previous comment, the `SAMPLE_PETS_feb152024` fix worked, thanks 💯

Also no worries about the temp contributor, was just checking I hadn't missed something 👍🏻
