+++ cart.py
@@ -22,6 +22,7 @@ from flask_session import Session  # cookie and other local info storage
 from tempfile import mkdtemp  # related to error on flask session and permissions
 from filelock import Timeout, FileLock  # https://py-filelock.readthedocs.io/en/latest/
 import glob, os  # handling all abstract files as pandas df, and file actions with os
+import os.path
 import json, signal # handling shutdown
 import pandas as pd  # for csv rows per grader
 import numpy as np  # data analysis
@@ -39,9 +40,11 @@ from ngrok_flask_cart import run_with_ngrok # https://pypi.org/project/ngrok-fla
 import shutil # for demo version of app 
 import signal # resetting option
 
-PATH_TO_ABSTRACTS = "abstracts/*.csv"
-DEFAULT_PATH = "abstracts/"
+
+PATH_TO_ABSTRACTS = os.path.join( os.path.dirname(os.path.realpath(__file__)), "abstracts", "*.csv" ) # "abstracts/*.csv"
+DEFAULT_PATH = os.path.join( os.path.dirname(os.path.realpath(__file__)), "abstracts", "" ) #"abstracts/"
 DEBUG = False
+
 
 
 ############################################# ------------------------------------file structure
@@ -125,7 +128,7 @@ app.secret_key = "*KRUD?zrAJ|jX=L]]y=[T)P@]<UEe'"
 
 Session(app)
 
-def start_lock(file_path):
+def start_lock(file_path, unique_id=None):
 
     if DEBUG:
         print("[-] starting lock process now")
@@ -133,6 +136,15 @@ def start_lock(file_path):
     file_path = file_path
     lock_path = file_path + ".lock"
     lock = FileLock(lock_path, timeout=timeout_for_locks)
+    if DEBUG: print(lock_path)
+    if DEBUG: print(file_path)
+    if os.path.isfile(lock_path) == False:
+        print("[-] GENERAL CALL: possibly windows machine, no .lock file created, use workaround by creating one")
+        os.chdir('abstracts')
+        with open(str(unique_id)+".lock", "w+") as f: f.close()
+        os.chdir('../')
+        if DEBUG: print("DONE windows patch\n\n")
+        if DEBUG: print(os.path.isfile(lock_path))
     try:
         with lock.acquire():
             temp_df = pd.read_csv(file_path, names=colnames, header=None)
@@ -190,6 +202,7 @@ def start_lock(file_path):
             )
         to_ret["success"] = False
         to_ret["lock"] = None
+    if DEBUG: print(os.path.isfile(lock_path))
     return to_ret
 
 
@@ -329,7 +342,7 @@ def index():
             session["user_counter"] = len(user_list)
 
             for f in all_files:
-                if os.path.isfile(f + ".lock"):
+                if os.path.isfile(f.replace(".csv",'') + ".lock"):
                     temp_df = pd.read_csv(
                         f, names=colnames, header=None
                     )  # bottleneck here
@@ -398,7 +411,7 @@ def index():
                 ):
                     if DEBUG:
                         print("----> found potential abstract, checking locks")
-                    ret = start_lock(f)
+                    ret = start_lock(f, current_id)
                     if ret["success"] == True:
                         if DEBUG:
                             print(ret)
@@ -416,10 +429,10 @@ def index():
                         parsed_time = moment.unix(parsed_time)
                         if DEBUG:
                             try:
-                                print(parsed_time.format("YYYY-M-D h:m A"))
+                                print(parsed_time.format("YYYY-M-D hh:m A"))
                             except:
                                 print("maybe error in computer interpreting 'A'")
-                                print(parsed_time.format("YYYY-M-D H:m"))
+                                print(parsed_time.format("YYYY-M-D HH:m"))
                         if DEBUG:
[…]
