+++ testing/test.py
@@ -22,6 +22,7 @@ from os import fdopen, remove
 from distutils.dir_util import copy_tree
 import glob 
 import threading
+import random 
 
 def isConnected():
     try:
@@ -81,6 +82,48 @@ def replace(file_path, pattern, subst):
     #Move new file
     move(abs_path, file_path)
 
+def clear_out_files():
+    os.chdir('../')
+    print("[ ] clear_out_files")
+
+    # clear out testing folder 
+    files = glob.glob('abstracts/*.csv')
+    for f in files:
+        print('removing! --> ', f)
+        os.remove(f)
+        time.sleep(.2)
+    files = glob.glob('abstracts/*.txt')
+    for f in files:
+        print('removing! --> ', f)
+        os.remove(f)
+    files = glob.glob('abstracts/*.lock')
+    for f in files:
+        print('removing! --> ',f)
+        os.remove(f)
+    try:
+        os.remove('abstracts/coders.txt')
+    except:
+        print("no coders.txt") 
+    try:
+        os.remove('abstracts/confetti.txt')
+    except:
+        print("no confetti.txt")
+    try:
+        os.remove('abstracts/ngrok_auth.txt')
+    except:
+        print("no ngrok_auth.txt") 
+    try:
+        os.remove('abstracts/ngrok_domain.txt')
+    except:
+        print("no ngrok_domain.txt")
+    try:
+        os.remove('abstracts/num_reviews_per_abstract.txt')
+    except:
+        print("no num_reviews_per_abstract.txt") 
+    print("[+] clear_out_files")
+    os.chdir('testing')
+    return
+
 def teardown():
     print("[ ] teardown")
 
@@ -89,8 +132,10 @@ def teardown():
     for f in files:
         print(f)
         os.remove(f)
-
-    os.remove("cart_testing.py")
+    try:
+        os.remove("cart_testing.py")
+    except:
+        print("no 'cart_testing.py' exists")
     print("[+] teardown")
     return
 
@@ -110,12 +155,15 @@ def test_vote(examples_to_use, number_to_vote):
             os.remove(f)
         
         # check that we have some files in the testing cateogry (copy over from -example_data_small)
+        print('[ ] grabbing files')
         files_to_copy = glob.glob('abstracts/' + examples_to_use + '/*')
         for f in files_to_copy:
             print(f)
-            shutil.copy(f, "abstracts/-testing/")
-        
-        # copy the file for testing 
+            shutil.copy(f, "abstracts/")
+        print('[+] done grabbing files')
+
+        # copy the file for testing
+        print('[ ] moving files into the right location for CART') 
         shutil.copyfile("cart.py", "cart_testing.py")
         file_path = 'cart_testing.py'
         pattern1 = 'PATH_TO_ABSTRACTS = "abstracts/*.csv"'
@@ -126,6 +174,7 @@ def test_vote(examples_to_use, number_to_vote):
         time.sleep(10)
         replace(file_path,pattern2,subst2)
         time.sleep(10)
+        print('[+] done moving files into the right location for CART') 
 
         print("[+] setup")
       
@@ -158,26 +207,34 @@ def test_vote(examples_to_use, number_to_vote):
         select = Select(browser.find_element('xpath', '/html/body/main/article/div/div/div/div/div/form/select'))
         # select by value 
         select.select_by_value('user1')
-        time.sleep(1)
+        time.sleep(10)
         actions = ActionChains(browser)
         actions.send_keys(Keys.ENTER)
         actions.perform()
         time.sleep(3) 
         print("[+] login DONE")
 
-        print("[ ] cast vote of yes on all abstracts in the small examples")
+        print("[ ] cast vote for all examples")
         # for each of the 5 abstracts, vote yes
[…]
