### 2024-06-28 #6494 @nathanReitinger
No worries, we've all been there!

Thanks for the notes, glad we can get some testing on Windows! Here's what I did to troubleshoot (tl;dr---pull the latest repo and use the windows version of the conda yml)

- used `Microsoft Windows [Version 10.0.22621.3737]` in vmware
- installed latest anaconda and made sure it was linked to my path (can type `conda` in a windows terminal and launch anaconda)
`(base) PS C:\Users\User\Desktop> conda --version`
`conda 24.5.0`
- used `pip` to install on windows
- created a yml for windows users (`cart-windows.yml`)

unrelated, I made sure that 'test.py' in the testing folder worked, which required me to make a few updates. I've pushed all of these now!
