+++ paper/paper.bib
@@ -11,7 +11,7 @@
 }
 
 @article{reitinger:2023,
-  title={Epsilon-differential privacy, and a twostep test for quantifying reidentification risk},
+  title={Epsilon-differential privacy, and a two step test for quantifying reidentification risk},
   author={Reitinger, Nathan and Deshpande, Amol},
   journal={Jurimetrics: The Journal of Law, Science \& Technology},
   volume={63},
@@ -77,7 +77,7 @@
 }
 
 @article{smit:2021,
-  title={Computer-Assisted Qualitative Data Analysis Software for Scoping Reviews: A Case of ATLAS.ti},
+  title={Computer-assisted qualitative data analysis software for scoping reviews: A case of ATLAS.ti},
   author={Smit, Brigitte and Scherman, Vanessa},
   journal={International Journal of Qualitative Methods},
   volume={20},
