+++ paper/paper.bib
@@ -11,7 +11,7 @@
 }
 
 @article{reitinger:2023,
-  title={Epsilon-differential privacy, and a two step test for quantifying reidentification risk},
+  title={Epsilon-differential privacy, and a two-step test for quantifying reidentification risk},
   author={Reitinger, Nathan and Deshpande, Amol},
   journal={Jurimetrics: The Journal of Law, Science \& Technology},
   volume={63},
