+++ paper/paper.md
@@ -30,7 +30,7 @@ bibliography: paper.bib
 
 # Statement of need
 
-A systematization of knowledge or systematic review paper is a popular type of paper found across multiple disciplines: medicine, computer science, sociology, psychology, economics, political science, marketing, and genetics. This type of paper most commonly involves a two-stage process. In the first stage, researchers gather a large set of papers—easily over 10,000—and then manually filter those papers (based on paper titles and abstracts) according to predefined relevancy criteria [@thomas:2021; @singhal:2023; @franz:2021; @reitinger:sok]. For example, imagine a researcher wanted to learn about legally-sufficient data sanitization tools [@bellovin:2019]. A paper on the prevalence of a particular web tracking technique like canvas fingerprinting [@reitinger:2021] would be marked as irrelevant, but a paper on statutory differentially privacy requirements [@reitinger:2023] would be marked as relevant. After this first step is complete—often resulting in a set of around 100 papers—most researchers then move on to analyzing papers in-depth, the second stage, using qualitative coding techniques well-fit by existing software.
+A systematization of knowledge or systematic review paper is a popular type of paper found across multiple disciplines: medicine, computer science, sociology, psychology, economics, political science, marketing, and genetics. This type of paper most commonly involves a two-stage process. In the first stage, researchers gather a large set of papers—easily over 10,000—and then manually filter those papers (based on paper titles and abstracts) according to predefined relevancy criteria [@thomas:2021; @singhal:2023; @franz:2021; @reitinger:sok]. For example, imagine a researcher wanted to learn about legally-sufficient data sanitization tools [@bellovin:2019]. A paper on the prevalence of a particular web tracking technique like canvas fingerprinting [@reitinger:2021] would be marked as irrelevant, but a paper on statutory differential privacy requirements [@reitinger:2023] would be marked as relevant. After this first step is complete—often resulting in a set of around 100 papers—most researchers then move on to analyzing papers in-depth, the second stage, using qualitative coding techniques well-fit by existing software.
 
 Surprisingly, software to support this first step of paper culling is lacking. Existing tools are either too generalist (resulting in ineffective paper reviewing) or too fine-tuned (resulting in tradeoffs impacting functionality). For example, generalist tools like Microsoft Excel or Google Sheets could allow teams of researchers to check the relevancy of snippets of text, but these tools can be clunky (i.e., given a scale of >10,000 papers in consideration) and not usable (i.e., reviewing hundreds of rows in a spreadsheet is error-prone and fatiguing). Other tools, like Atlas.ti [@smit:2021] or maxQDA [@Franzosi:2012], could allow teams of researchers to "code" for relevancy, but do not naturally accommodate dynamic participation (e.g., team members might not be able to review the same amount of papers) or variable review rates (e.g., sometimes you want all team members to review all papers, and sometimes you want at least two, but you do not set who those team members are). Enter `CART`.
 
