+++ paper/paper.md
@@ -24,15 +24,15 @@ bibliography: paper.bib
 
 # Summary
 
-`CART` (Culling Abstacts for Relevancy in Teams) provides a way for a team of individuals to complete the first step common to most systematization of knowledge (SoK) projects[^1]: culling, in teams, a large set of research papers (i.e., titles and abstracts) based on per-project relevancy criteria. Existing software solutions either do not fit this need in a usable way—leading to error-prone voting—or do not offer the necessary functionality to work on this problem effectively.
+`CART` (Culling Abstacts for Relevancy in Teams) provides a way for a team of individuals to complete the first step common to most systematization of knowledge (SoK) projects[^1]: culling, in teams, a large set of research papers (i.e., titles and abstracts) based on per-project relevancy criteria. Existing software solutions either do not fit this need in a usable way—leading to error-prone voting—or do not offer the necessary functionality to work on the problem effectively.
 
 `CART` provides both functionality and usability: *Functionally*, `CART` accommodates teams of any size (i.e., accommodating race conditions), review requirements of any number (e.g., mandating that each paper is reviewed by *n* team members), and dynamic participation rates (e.g., team members may participate in as little or as much reviewing as they like). *Usability-wise*, `CART` provides a gamified user interface, making the process of reviewing papers easier, and therefore more accurate and efficient. 
 
-# Statement of need
+# Statement of Need
 
-A systematization of knowledge or systematic review paper is a popular type of paper found across multiple disciplines: medicine, computer science, sociology, psychology, economics, political science, marketing, and genetics. This type of paper most commonly involves a two-stage process. In the first stage, researchers gather a large set of papers—easily over 10,000—and then manually filter those papers (based on paper titles and abstracts) according to predefined relevancy criteria [@thomas:2021; @singhal:2023; @franz:2021; @reitinger:sok]. For example, imagine a researcher wanted to learn about legally-sufficient data sanitization tools [@bellovin:2019]. A paper on the prevalence of a particular web tracking technique like canvas fingerprinting [@reitinger:2021] would be marked as irrelevant, but a paper on statutory differential privacy requirements [@reitinger:2023] would be marked as relevant. After this first step is complete—often resulting in a set of around 100 papers—most researchers then move on to analyzing papers in-depth, the second stage, using qualitative coding techniques well-fit by existing software.
+A systematization of knowledge or systematic review paper is a popular type of paper found across multiple disciplines: medicine, computer science, sociology, psychology, economics, political science, marketing, and genetics. This type of paper most commonly involves a two-stage process. In the first stage, researchers gather a large set of papers—easily over 10,000—and then manually filter those papers (based on paper titles and abstracts) according to predefined relevancy criteria [@thomas:2021; @singhal:2023; @franz:2021; @reitinger:sok]. For example, imagine a researcher who wanted to learn about legally-sufficient data sanitization tools [@bellovin:2019]. A paper on the prevalence of a particular web tracking technique like canvas fingerprinting [@reitinger:2021] would be marked as *irrelevant*, but a paper on statutory differential privacy requirements [@reitinger:2023] would be marked as *relevant*. After this first step is complete—often resulting in a set of around 100 papers—most researchers then move on to analyzing papers in-depth, the second stage, using qualitative coding techniques well-fit by existing software.
 
[…]
