### 2024-07-15 #6494 @Nikoleta-v3
> Should I mark these somehow in the Bib or should we ignore the errors?

We can just ignore the error 😄

@nathanReitinger one question: the name of the package in `CART` for Culling Abstacts for Relevancy in Teams. Should the second word be `abstracts`?

Moreover, could you please:

- (If necessary) Make a new tagged release of your software and list the version tag of the archived version here. Alternatively, let me know if I should just use release v1.0.0.
- Archive the reviewed software in Zenodo or a similar service (e.g., Figshare, an institutional repository).
- Ensure the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (which should match the paper title) and the author list (make sure the list is correct and that people who only made small fixes are not included). You may also add the authors' ORCID IDs.
- List the DOI of the archived version here.
