### 2024-07-15 #6494 @nathanReitinger
What a great catch @Nikoleta-v3 👍👍👍THANK YOU :)

- tagged release: `v1.0.01` https://github.com/nathanReitinger/CART/releases/tag/v1.0.0
- Figshare: https://figshare.com/articles/software/CART_Culling_Abstracts_for_Relevancy_in_Teams/26304985
- Figshare DOI: https://doi.org/10.6084/m9.figshare.26304985
