### 2024-07-15 #6494 @Nikoleta-v3
> What a great catch @Nikoleta-v3 👍👍👍THANK YOU :)

🎉 @nathanReitinger thank you for your submission, and again, apologies for the delay in the process.

@jhculb @atanikan, thank you both very much for your time in reviewing for JOSS! 🙏🏻
