### 2024-07-15 #6494 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
