### 2024-07-16 #6494 @danielskatz
Congratulations to @nathanReitinger (Nathan Reitinger) on your publication!!

JOSS depends on volunteers and wouldn't be able to succeed without you
